Moved logwriter to utils root
