+++ clinicadl/utils/maps_manager/__init__.py
@@ -1,8 +1,8 @@
+from ..logwriter import LogWriter
 from .iotools import (
     check_and_complete,
     commandline_to_json,
     read_json,
     write_requirements_version,
 )
-from .logwriter import LogWriter
 from .maps_manager import MapsManager
